backend/requirements.txt
# --- FastAPI Framework ---
fastapi
uvicorn
python-dotenv
python-multipart
pydantic-settings
email-validator

# --- Database & Storage (Supabase) ---
supabase
requests
httpx

# --- Image Processing ---
# pillow-heifはiPhoneのHEIC形式を扱うために必須
Pillow
pillow-heif

# --- AI Core (PyTorch CPU) ---
# 依存関係の競合を防ぐため、torch/torchvisionをセットで指定
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
numpy

# --- OpenAI CLIP (GeoCLIP Backbone) ---
# CLIPのテキスト処理・正規化に必要
ftfy
regex
git+https://github.com/openai/CLIP.git

# --- GeoCLIP & Utilities ---
# 進捗表示用
tqdm
# 座標から住所への変換
geopy
# GeoCLIP本体
geoclip